Refactored the package so it's edu.calvin.cs262.scrumptious
